% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{register_completion}
\alias{register_completion}
\title{Register completion functions}
\usage{
register_completion(...)
}
\arguments{
\item{...}{One or more completion functions specified as named parameters.}
}
\description{
Completion functions should take one parameter `env`, the completion
environment, see `?rc.settings` for details of this environment. They should
simply return any completions found or `return(NULL)` otherwise.
}
\details{
If all registered completions do not have any completions for a given
context, than R's standard completions are used.
}
