% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{return_unless}
\alias{return_unless}
\title{Return from a function unless a condition is met}
\usage{
return_unless(x)
}
\arguments{
\item{x}{Condition to be evaluated.}
}
\description{
This function can be used to return early from a function unless a condition
is met. It is equivalent to `if (!x) return()`.
}
