% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{helpers}
\alias{helpers}
\alias{current_function}
\alias{inside_quotes}
\alias{is_first_argument}
\alias{current_argument}
\title{Completion helpers}
\usage{
current_function(env)

inside_quotes(env)

is_first_argument(env)

current_argument(env)
}
\arguments{
\item{env}{The completion environment, see `?rc.status()` for details.}
}
\description{
Completion helpers
}
\section{Functions}{
\itemize{
\item \code{current_function}: Returns the current function call, or `""` if
not within a call.

\item \code{inside_quotes}: Returns `TRUE` if within single or double quotes.

\item \code{is_first_argument}: Returns `TRUE` if currently completing the first argument.

\item \code{current_argument}: Returns the current named argument, or `""` if not
completing a named argument.
}}

