% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{helpers}
\alias{helpers}
\alias{inside_quotes}
\alias{is_comment}
\alias{is_roxygen_comment}
\alias{inFunction}
\title{Completion helpers}
\usage{
inside_quotes(env)

is_comment(env)

is_roxygen_comment(env)

inFunction(env)
}
\arguments{
\item{env}{The completion environment, see `?rc.status()` for details.}
}
\description{
Completion helpers

originally Adapted from utils:::inFunction()
}
\section{Functions}{
\itemize{
\item \code{inside_quotes}: Returns `TRUE` if within single or double quotes.

\item \code{is_comment}: Returns `TRUE` if the current context is a comment.

\item \code{is_roxygen_comment}: Returns `TRUE` if the current context is a roxygen comment.

\item \code{inFunction}: Checks to see if we are in a function. If
we are then it returns an list loaded with information about what has been typed so far
}}

