% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{register_completion}
\alias{register_completion}
\title{Register completion functions}
\usage{
register_completion(...)
}
\arguments{
\item{...}{One or more completion functions specified as named parameters.}
}
\description{
Completion functions should take one parameter `env`, the completion
environment, see `?rc.settings` for details of this environment. They should
set the `comps` and `fileName` fields of the completion and return `TRUE` if
no further completions should be attempted. They should return `FALSE` if
there are no completions for the current context.
}
\details{
If all registered completions return `FALSE` for a given context, than R's
standard completions are used.
}
